Changed addsamplestobackground default value to T
Adding samples to the background data is now the default.  This eliminates the possibility of the model being infeasible.  If presence samples are not added to the background data, infeasibility can cause problems with categorical features (for example, if the presence samples contain a category that isn't present in the background data) and hinge features (for example, if the presence samples contain a value for some variable that is outside that variable's range in the background data).  Infeasibility causes glmnet to stop early, with more regularization than requested by maxnet.